+++ src/components/Platform/03Inventories/Merchandises/04ModalEditMerchandise/ModalEditMerchandise.tsx
@@ -91,7 +91,7 @@ function ModalEditMerchandise({ token, idItem, merchandise, branches, onCloseMod
     return (
         <div>
             <div className={`${styles.containerCard} m-auto d-flex flex-column align-items-center justify-content-center`}>
-                <h1 className={`${styles.title} text-center`}>Información de la mercancía</h1>
+                <h1 className={`${styles.title} text-center`}>Edita la información de la mercancía</h1>
             </div>
 
             <div className="w-100">
