+++ src/components/Platform/03Inventories/Merchandises/04ModalEditMerchandise/styles.module.css
@@ -3,8 +3,6 @@
 }
 .title {
     color: #f28705;
-    font-size: 50px;
-    font-weight: 600;
     letter-spacing: 1px;
 }
 .label {
