+++ src/pages/Platform/PanelUser/03Inventories/03InventoryProducts/ConsultProducts/ConsultProductsPage.tsx
@@ -199,7 +199,7 @@ function ConsultProductsPage() {
                                             </div>
 
                                             <div className={`${styles.selling__Price} pt-0 pb-0 px-2 d-flex align-items-center justify-content-center overflow-hidden`}>
-                                                <span className={`${styles.text__Ellipsis} overflow-hidden`}>$ {product.IVA} %</span>
+                                                <span className={`${styles.text__Ellipsis} overflow-hidden`}>{product.IVA} %</span>
                                             </div>
                                             <div className={`${styles.selling__Price} pt-0 pb-0 px-2 d-flex align-items-center justify-content-center overflow-hidden`}>
                                                 <span className={`${styles.text__Ellipsis} overflow-hidden`}>$ {formatNumber(product.sellingPrice)}</span>
